V4.12.1.180 - Increased range of alignments to cope with LPC 256-byte aligned programming

+++ DeviceDatabase_DLL/src/DeviceData.cpp
@@ -586,7 +586,7 @@ AddressType MemoryRegion::getAddressType() const {
 }
 
 // Get memory region data alignment (1,2,4 etc)
-uint8_t     MemoryRegion::getAlignment() const {
+uint32_t     MemoryRegion::getAlignment() const {
    return alignment;
 }
